requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
transformers>=4.45.0
accelerate>=0.25.0
peft>=0.7.0
pillow>=10.0.0
qwen-vl-utils>=0.0.8
